+++ examples/forgetful.rs
@@ -1,10 +1,15 @@
-use std::collections::HashMap;
+use std::{collections::HashMap, time::Duration};
 
 use espionox::{
-    agents::{memory::Message, Agent},
+    agents::{
+        memory::{Message, MessageStack},
+        Agent,
+    },
     environment::{
         agent_handle::{EndpointCompletionHandler, MessageRole},
-        dispatch::{listeners::ListenerMethodReturn, Dispatch, EnvListener, EnvMessage},
+        dispatch::{
+            listeners::ListenerMethodReturn, Dispatch, EnvListener, EnvMessage, EnvRequest,
+        },
         Environment,
     },
     language_models::{
@@ -12,6 +17,7 @@ use espionox::{
         ModelProvider,
     },
 };
+use tokio::time::sleep;
 
 #[derive(Debug)]
 pub struct Forgetful {
@@ -27,20 +33,14 @@ impl From<&str> for Forgetful {
 
 impl<H: EndpointCompletionHandler> EnvListener<H> for Forgetful {
     fn trigger<'l>(&self, env_message: &'l EnvMessage) -> Option<&'l EnvMessage> {
-        match env_message {
-            EnvMessage::Response(noti) => {
-                if let Some(id) = noti.agent_id() {
-                    if id == &self.watched_agent_id {
-                        Some(env_message)
-                    } else {
-                        None
-                    }
-                } else {
-                    None
+        if let EnvMessage::Request(req) = env_message {
+            if let EnvRequest::GetCompletion { agent_id, .. } = req {
+                if agent_id == &self.watched_agent_id {
+                    return Some(env_message);
                 }
             }
-            _ => None,
         }
+        None
     }
 
     fn method<'l>(
@@ -75,22 +75,25 @@ async fn main() {
         .unwrap();
 
     let fgt = Forgetful::from("jerry");
-    env.insert_listener(fgt).await;
-    env.spawn().await.unwrap();
+    let _ = env.insert_listener(fgt).await;
+    let mut env_handle = env.spawn_handle().unwrap();
     let message = Message::new_user("whats up jerry");
     for _ in 0..=5 {
         let _ = jerry_handle
             .request_io_completion(message.clone())
             .await
             .unwrap();
+        sleep(Duration::from_millis(200)).await;
     }
-    env.finalize_dispatch().await.unwrap();
-    let dispatch = env.dispatch.write().await;
+    let state_ticket = jerry_handle.request_state().await.unwrap();
+    let mut stack = env_handle.finish_current_job().await.unwrap();
+    let noti: espionox::environment::dispatch::EnvNotification =
+        stack.take_by_ticket(state_ticket).unwrap();
 
-    let jerry = dispatch.get_agent_ref(&jerry_handle.id).unwrap();
+    let jerry_m_stack: &MessageStack = noti.extract_body().try_into().unwrap();
 
-    println!("Jerry stack: {:?}", jerry.cache);
-
-    assert_eq!(jerry.cache.len(), 0);
+    println!("Jerry stack: {:?}", jerry_m_stack);
+    let stack_sans_system = jerry_m_stack.ref_filter_by(MessageRole::System, false);
+    assert_eq!(stack_sans_system.len(), 0);
     println!("All asserts passed, forgetful working as expected");
 }