+++ examples/summarize_at_limit.rs
@@ -44,7 +44,9 @@ impl<H: EndpointCompletionHandler> EnvListener<H> for SummarizeAtLimit<H> {
                 agent_id, cache, ..
             } = noti
             {
-                if cache.len() >= self.limit && agent_id == &self.watched_agent_id {
+                if cache.ref_filter_by(MessageRole::System, false).len() >= self.limit
+                    && agent_id == &self.watched_agent_id
+                {
                     return Some(env_message);
                 }
             }
@@ -77,7 +79,7 @@ impl<H: EndpointCompletionHandler> EnvListener<H> for SummarizeAtLimit<H> {
                 .get_agent_mut(&self.watched_agent_id)
                 .expect("Failed to get watched agent");
             watched_agent.cache.mut_filter_by(MessageRole::System, true);
-            watched_agent.cache.push(Message::new_system(&summary));
+            watched_agent.cache.push(Message::new_assistant(&summary));
             Ok(trigger_message)
         })
     }
@@ -106,27 +108,31 @@ async fn main() {
         .unwrap();
     let sal = SummarizeAtLimit::new(5usize, "jerry", summarizer);
 
-    env.insert_listener(sal).await;
-    env.spawn().await.unwrap();
+    env.insert_listener(sal).await.unwrap();
+    let mut env_handle = env.spawn_handle().unwrap();
+
     for _ in 0..=5 {
         jerry_handle
             .request_cache_push(
                 "im saying things to fill space".to_owned(),
-                MessageRole::System,
+                MessageRole::User,
             )
             .await
             .expect("failed to request cache push");
         tokio::time::sleep(Duration::from_millis(200)).await;
     }
-    env.finalize_dispatch().await.unwrap();
-    let stack = env.notifications.0.write().await;
-    let messages = match stack.get(0).unwrap() {
-        EnvNotification::AgentStateUpdate { cache, .. } => cache.as_ref(),
-        _ => panic!("First on stack should be a cache update"),
-    };
+    let mut stack = env_handle.finish_current_job().await.unwrap();
+    let latest = stack.pop_back().unwrap();
 
-    assert_eq!(messages.len(), 2);
-    assert_eq!(messages[0].role, MessageRole::System);
-    assert_eq!(messages[1].role, MessageRole::User);
-    println!("All asserts passed, summarize at limit working as expected");
+    // env.finalize_dispatch().await.unwrap();
+    if let EnvNotification::AgentStateUpdate { cache, .. } = latest {
+        println!("STACK: {:?}", cache);
+        assert_eq!(cache.len(), 3);
+        assert_eq!(cache.as_ref()[0].role, MessageRole::System);
+        assert_eq!(cache.as_ref()[1].role, MessageRole::Assistant);
+        assert_eq!(cache.as_ref()[2].role, MessageRole::User);
+        println!("All asserts passed, summarize at limit working as expected");
+        return;
+    }
+    println!("Incorrect notification in last place: {:?}", latest);
 }