+++ src/environment/errors.rs
@@ -7,6 +7,16 @@ use std::fmt::{Debug, Display, Formatter};
 pub use super::dispatch::listeners::error::ListenerError;
 
 #[derive(thiserror::Error)]
+pub enum EnvHandleError {
+    CouldNotOwnNotifications,
+    MissingNotifications,
+    MissingHandleData,
+    MissingThreadHandle,
+    ThreadAlreadySpawned,
+    EnvError(#[from] EnvError),
+}
+
+#[derive(thiserror::Error)]
 pub enum EnvError {
     #[error(transparent)]
     Undefined(#[from] anyhow::Error),
@@ -14,7 +24,9 @@ pub enum EnvError {
     Dispatch(#[from] DispatchError),
     Join(#[from] tokio::task::JoinError),
     Timeout(#[from] tokio::time::error::Elapsed),
+    MissingHandleData,
     Request(String),
+
     Send,
 }
 
@@ -42,6 +54,18 @@ impl Debug for EnvError {
 }
 
 impl Display for EnvError {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{:?}", self)
+    }
+}
+
+impl Debug for EnvHandleError {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        error_chain_fmt(self, f)
+    }
+}
+
+impl Display for EnvHandleError {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         write!(f, "{:?}", self)
     }