+++ src/language_models/anthropic/mod.rs
@@ -2,7 +2,7 @@ use reqwest::header::HeaderMap;
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 
-use crate::environment::agent_handle::{MessageRole, MessageStack};
+use crate::environment::agent_handle::{Message, MessageRole, MessageStack};
 
 use super::{endpoint_completions::EndpointCompletionHandler, error::ModelEndpointError};
 
@@ -46,6 +46,34 @@ impl EndpointCompletionHandler for AnthropicCompletionHandler {
         "https://api.anthropic.com/v1/messages"
     }
 
+    fn agent_cache_to_json(cache: &MessageStack) -> Vec<Value> {
+        // Anthropic model requires that messages alternate from User to assistant. So we'll
+        // concatenate all adjacent messages to one
+        let mut val_vec: Vec<Value> = vec![];
+        let mut last_message: Option<Message> = None;
+
+        for message in cache.clone().into_iter() {
+            match last_message.take() {
+                Some(mut m) => {
+                    if message.role == m.role {
+                        m.content = format!("{}. {}", m.content, message.content);
+                        last_message = Some(m);
+                    } else {
+                        let val: Value = m.into();
+                        val_vec.push(val);
+                        last_message = Some(message);
+                    }
+                }
+                None => last_message = Some(message),
+            }
+        }
+        if let Some(m) = last_message {
+            let val: Value = m.into();
+            val_vec.push(val);
+        }
+        val_vec
+    }
+
     fn request_headers(&self, api_key: &str) -> HeaderMap {
         let mut map = HeaderMap::new();
         map.insert("x-api-key", format!("{}", api_key).parse().unwrap());
@@ -58,7 +86,7 @@ impl EndpointCompletionHandler for AnthropicCompletionHandler {
         let system_stack: MessageStack = messages.ref_filter_by(MessageRole::System, true).into();
         let sans_system_stack: MessageStack =
             messages.ref_filter_by(MessageRole::System, false).into();
-        let context: Vec<Value> = (&sans_system_stack).into();
+        let context = Self::agent_cache_to_json(&sans_system_stack);
         let system_content = system_stack
             .as_ref()
             .into_iter()
@@ -75,7 +103,7 @@ impl EndpointCompletionHandler for AnthropicCompletionHandler {
     }
 
     fn handle_io_response(&self, response: Value) -> Result<String, ModelEndpointError> {
-        println!("value: {:?}", response);
+        // println!("value: {:?}", response);
         let response = AnthropicResponse::try_from(response).unwrap();
         Ok(response.content[0].text.to_owned())
     }